chore: reorganized framework module locations
